UX3/requirements.txt
flask
python-dotenv
google-generativeai
APScheduler
